Not sending comments when commit is not head
As we described on https://github.com/codecov/worker/pull/549 , there are times where the information
we have about a pull request doesn't match the information that Github has (which is the source of truth and is the more recent one)

But we recently found race conditions (not that uncommon) where PR comments could be triggered when
even our DB already knows that commit is not the PR head.

On those cases, sending a PR with a warning might be even worse. Since on those cases our
system already updated everything, there is a reasonable chance the proper PR comment, with the proper head,
was already sent, and sending a new one will overwrite the right one, and might not get overwritten later

And with that warning on top, this might be problematic

So this PR aims to:

1) Fix some cases where PR head might be wrongly assigned (due to race conditions)
2) Avoid PR comments when we know that we already have data from more recent commits.