Refactoring at which point of the stack processing happens
Moving some of the "heavy-duty" processing of report from `UploadProcessorTask.do_process_individual_report` to `ReportService.build_report_from_raw_content`.

This second function was already a function that was called. But now we are delegating more of the logic to it so it can return a `ProcessingResult` object with all the relevant information about the result of the processing, in a self-contained unit.

The goal here is to make the processing something more "repeatable", where we can just take the `Upload` object and run a function to it to see what the results would be.

It also frees us to new ways of processing reports in the tasks. It could, for example, make it easier to save each `ProcessingResult` separately and join them all at a later moment in one final report result.