Making percent_type schema nullables
It looks like that `voluptuous`, differently from `schema`, errors that do not extend from Invalid are not ingested by the validator as errors.

So it leads to cases where a `TypeError` exception inside the validator just crashes the whole call.

On this case, the fix is to just accept `None` on the values so the `percent_type` validation method is not called with `None` as an argument.