Changing when we are saving the report totals
As it is known, on the bash uploader we load multiple upload-files (as in coverage.json, coverage.xml etc) on the same upload.

As it turns out, if an upload contains N upload-files, the `session.totals` is being assigned N times, each time the value of each of those report.totals. Meaning session.totals is never the right value: the totals of the report merging all those files. It instead gets the value of the last non-empty processed report for that upload

Unfortunately, to fix that, we need a report where there is nothing but the results of those upload processing. So we need to set a temporary "holder report", and have it hold all the results from this upload, merging one after the other.

Then, after all is done, the `temporary_report.totals` value is assigned to the `session.totals`. And then merged in the final report

The hard part is that, despite this looking correct, the `merge` method is not as clean and transparent as it looks like. So we have to make sure, on testing, that we did not break any case.

Also changing some tests on joined. Still not sure how that feature is being used currently, but the tests are still reasonable