Experimenting with moving tests to httpx
Consistently building tests that are both easily modifiable and secure is becoming harder with codecov_vcr, given that it needs credentials if we want to build new tests.

For some personal tokens this should be fine, but the more "this needs real complex data" tests, it might be better to start relying on `respx` (https://lundberg.github.io/respx/ ) tests.

This attempts to write down how those tests would look