Solving the worker side of statementMap errors
Making it so we don't just crash when `statementMap` is not present in the dict.

We are doing exactly what the same Node processor already does at a different point and what seems to be the original intention of the design: if something is not a dict (or is just missing) treat it like an empty dict, which essentially makes us skip the information we would have gathered from that section. Another piece of the code already does exactly `data.get("statementMap")`

In this case, maybe the customer doesn't have a `statementMap` on their report, but they might have a `branchMap` or a `fnMap`. So even without a statementMap, they might still get some useful information. So not having a `statementMap` technically does not make this a `CorruptRawReportError` (like I considered doing).

Even without any of those fields, it could be that the report has `lineData` or `linesCovered`, which does make the report a not-corrupt report.

But the sentry data shows us that on most cases, there is just nothing there. Usually when one field is missing, they all are.

The deeper issue here is that, since legacy, node is the `catch-all` processor. Any json report that does not fit any of the other reports just ends up caught by the NodeProcessor.

It would just be nice for someone eventually go in and determine "Those are the rules that make a report a node report" so we can implement them and make sure other files don't wrongly try to get processed here

So this solution implemented here does not put the user in an ideal situation. They will very likely get an empty report error now. But at least empty reports are better shown in the UI, and this stops filling up our sentry

The user will need to upload different reports (which it looks like a relevant ticket is already working on)