Adding Ignore Coverage
We found that in production, there are lines like

```
[null, "b", [[157, null]]]
```

Basically, there are processors that just return `null` as coverage. And python treats that as a line to be ignored.

So we need to ignore it.

Originally, we were just taking lines that started with null and marking as empty lines. But then, if the above line is valid, so is the one below:

```
[1, "b", [[157, null], [1, 1]], null, null]
```

So the empty coverage is not only possible on the start. It might happen in the middle and not on the start. And on the line above, general coverage is Hit, coverage of session-1 is Hit, but the session-157 coverage is "skip this line".

So we just have to treat this as a Coverage possible value, and make sure everywhere in this code this line is a noop