Dealing with a conflict on the database
We have in the repos table two constraints:

```
create unique index repos_slug on repos (ownerid, name);

create unique index repos_service_ids on repos (ownerid, service_id);
```

They are both sensible constraints.

But they are two sides of the same idea of uniqueness. And sometimes they might conflict with each other.

For example, in the SyncReposTask we have a behavior that doesn't play well with them: the moving of repos.

It is unclear what situation leads to this, but the (most likely) situation is like this:

There was a repo abc on both owners albert and otto. So one repo `albert/abc` and one repo `otto/abc`.

`albert/abc` has service_id 123
`otto/abc` has service_id 456

Then albert deleted its own repo, and otto moved its own repo to albert ownership.

Now, we have a repo `albert/abc` with service_id 456 (service_ids don't change on moves)

From github perpective, this is all fine. The thing is our system.

On syncing "repo `albert/abc` with service_id 456", our system looks at our database sees that there is already a repo under albert (right username) but wrong service_id (123).

It also sees another repo with the right service_id (456) but under otto (wrong username)

The bug in here is that the system was trying to take the repo with service_id 456 and move it to be under `albert`. What simply was denied by the constraints, because there is already a repo named `abc` under albert.

What this fix does is try to identify this situation, and REFUSE TO DO ANYTHING. We just return the repoid with the right service_id (because that's the source of truth on identifiers), and not try to change anything.

This just ensures the rest of the syncing is not broken by this

The ideal solution will require a decision from product, because the right thing IS to move the moved repo to the new place, but first we need to either delete the old repo or rename it to something weird so the conflict doesn't happen. It does seem clear that the repo with service_id 123 is not accessible in any way on Github, so it might need to be deleted on our end.

Also I replaced the nested ifs with a bunch of early returns so the code flow was a bit more readable. This did hurt indentation.