Enabling encryption rolling keys
The main idea here is that now the whole encryption keys picking logic will be done from here. Not from the worker or the API

The "key creation" logic is done at `get_encryptor_from_configuration`, which gives a mapping to `StandardEncryptor` so it picks the right encryptors for the right tasks

First of all, if you don't know, we make an encryptor by concatenating a bunch of given secret strings in a special way to make a key

Here is how it works (if you change something, please make the effort to update the docs):

The customer will have, in their INSTALL YAML, the following:

```
    setup:
        encryption_secret: <their_legacy_encryption_secret"
        encryption:
            keys:
                - code: abc
                  value: "higsgilasifgshrvwa_45h3!@#$$@#ds"
                - code: jki
                  value: "bqwbdwagydics_49s"
                - code: plo
                  value: "uhsfsaigaernjchu3987$565*w"
            write_key: jki
```

Notice our `new_key_style` variable in the code. That's the place we can put new hardcoded keys if we ever need to reroll them again.

Assume for a small moment that we have multiple hardcoded keys: v1, v2, v3, v4, v5

Then our system will cross all those hardcoded keys with all the user-given keys. And will produce a new Encryptor with each pair. Meaning we will get 15 Encryptors:

```
    - (v1, abc) - by using the hardcoded key v1 and user-given key abc
    - (v1, jki) - by using the hardcoded key v1 and user-given key jki
    - (v1, plo)
    - (v2, abc)
    - (v2, jki)
    ...
    - (v4, plo)
    - (v5, abc)
    - (v5, jki)
    - (v5, plo)
```

So, they key (v1, abc) will get a "identifier" v1_abc. Each identifier gets an identifier on a similar fashion. So, everytime a specific Encyptor is used to generate a secret, we will prepend its identifier to the secret. So the final value of encoding a string with (v3, jki) will be "v3_jki::<somesecret>"

With those values prepended, it's easy later to know which encryptor was used to generate what. So, all we have to do is to pick the encryptor with the same identifier as the encoded key gives.

All the user has to do is to not change a key value after it is put in production. The same way they already can't change encryption_secret

Then, our system will produce a special Encryptor for the legacy key (which, notice, is the only one that uses the envvar ENCRYPTION_SECRET). Everytime an encoded string arrive without an identifier, we will know it's a legacy one, and pick the legacy generator for it.

One very important thing to notice here is that we need to keep those things compatible. Meaning:

1) Customers upgrading to this version without changing their install YAML need it not break anything
2) During the upgrade window, this need to play well with other containers still in the old version of the cod