Properly marking fraction as partials
In a subset of the Node Processor cases (ie, people with only one upload, and only on precalculated totals), it seems that we improperly mark some partials as hits.

It doesn't seem to affect a lot of people.

The fix is to just properly determine if the fraction is less than 1