Allowing empty layouts
We currently allow customers to pick any layout structure they want. This allows them to customize how they want their comment.

But due to a bug, we don't accept empty strings, meaning customers can't pass an empty layout.

They are able to pass a `null` to achieve that goal, which is a decent workaround. But it should also work with empty strings

So we are just removing the empty-string results from the "set-comparison" test we use on the layout validation