Lazy loading rust
Starting to have rust report lazy loaded

This is one of the first attempts to have rust lazy loaded on the system.

We believe that the situations in which rust is useful are a small number. But that on those cases, which are more common in heavy usage customers, they make a big difference.

So we want to make sure to limit the rust loading only on the cases they are actually useful. Otherwise, they will require more CPU than usual and not be use at all.

One benefit here is that by using `has_precalculated_totals`, we can control and select more cases where we know the python report will do a good enough job of processing totals. Such as when the session.totals is already pre-calculated and we know the filter can use those (which is being done on a separate set of PRs).

This way, we intend to load rust only whenever it will be impactful, without wasting resources.