Fixing doubling logging some cases
On our analysis we are using both `Rust was worse than python on the metrics calculation` and `Rust saved more than` logs. But on this "accumulated timers" strategy, we might end up logging some timers twice, since we do not clear the timers after the `Rust was worse than python on the metrics calculation` case.

This is fixing that by putting them on the same log line