Fixing first line logging
On `File format could not be recognized` we log the first line, so it's easier for us to debug the problem.

But there are no guarantees about the size of the first line. It could be very well be the whole file content

So we are limiting it to the first 100 chars.