[CODE-291] Adding configurable upload processing delay
Making it so the user can set

```
"setup" -> "upload_processing_delay"
```

And this way, if many Upload tasks start popping at a very small interval of time, the UploadTasks can wait until there is enough time passed with no tasks. This kind of guarantees that the user already sent all the relevant uploads for that "upload wave".

We still have to set this on the UploadHandler on legacy and API. But once we do, this will make it so more tasks are concentrated into the same chain without needing to rely a lot on retries.