Making torngit exceptions picklable
We were wrongly trying to pass "hardcoded" parameters when initializing the super() class of an exception. That does not work well with pickle for Exceptions because they rely on the actual parameters send on the super().__init__().

So we are stopping messing with those

Also created one `TorngitClientGeneralError` to have the concrete exceptions of `TorngitClientError`, so that one can only be used to catch the exceptions and not interfere on implementation of the exception classes