Adding installation config validator
Adding some validation to the config. This aims to make enterprise customers life a bit easier if their application doesn't work as expected.

This validation will try to do something but will not enforce the validation, because we don't want to block the customer yet if we make a mistake.

Also adding a particular feature that will be specially useful to both those customers and us:

Currently, a customer can set envvars as

```
ONE__TWO__THREE='value'
```

, and we will properly put it inside the config as:

```
{ 'one': { 'two': { 'three': 'value' }}}
```

But envvars are limited to strings. So customers don't have an easy way of passing booleans, arrays and dicts without us explicitly casting that.

So we are adding the `JSONCONFIG___` prefix. Every time an envvar starts with `JSONCONFIG___`, we parse that value as a JSON and use that instead. This allows the customer for example to do:

```
JSONCONFIG___SETUP__MEDIA='{"assets": "aaa", "dependancies": "bbb"}'
```

and we will properly set

```
{'setup': {'media': {"assets": "aaa", "dependancies": "bbb"}}}
```

Also removing `validate_yaml` from `get_config`

Since that is already validated now