Falling back to not student when app is not capable
The `https://education.github.com/api/user` endpoint only works when the app (as in Github Integration) is previously whitelisted by Github. Which happens in a very manual way on their end.

We don't want to break login when that happens. So we are just marking those users as not students