Setting lock times based on task time limit
We currently set lock times to a hardcoded value, based on what we have in production.

But enterprise customers might have set higher limits. So we need to accommodate the lock time so it doesn't expire before the task is finished. So we need to set it to a higher value than what the task can possibly take: the hard time limit