Trying multiple bases
In order to try the multiple bases given by Cobertura, we will be using the structure already existing on `BasePathAwarePathFixer` of `bases_to_try`. That will allow us to conditionally consider the extra base paths if the path-fixing conditions allow (ie, we have a TOC)

We now should be able to attempt multiple bases on a cobertura report, until one of them works at least. Otherwise, it will return None (which it already would since the original path failed)