Using initialization script
Unifying enterprise and production by using the same python script for initializing both.

The initial goal here was to avoid having initialization code run at import time (like the snippet of `initialize_sentry` does on app.py).

The best way to do that is to do it at the same place where we already do some initialization for enterprise. But for that we need for production to also use the same initialization code as enterprise.

I don't expect any performance changes or anything like that, since legacy has a similar script, and it looks like this essentially calls the same thing as the direct celery command already calls. It also helps us pick a single point where "initialization code" happens

A bunch of more tests are also being added