Centralizing get_config on the shared one
There seems to be now many implementations of `get_config`. As we implement new features on the one on `shared`, those are diverging.

This means that the user will have different results depending on what `get_config` the code is calling at a given point.

So we are changing all of them to call the same `shared` one.

It doesn't look like the default_configs (that re diverging from one another, by the way) have any crucial default that the `shared` one does not have ( https://github.com/codecov/shared/blob/master/shared/config/__init__.py#L20 ). But happy to keep if something is important