Fixing some test flakyness
We currently have constraints on some dbs that block us from repeating the same owner username on different rows

This is made problematic if our username creation comes from a randomized generator. This makes it so the same username appearing twice be possible but not usual. So this leads to flaky tests.

One of the solution is to use names that are different, but not random. Such as deterministic names that are generated from the test name.

This is what we are doing here by using the `request` fixture from pytest: https://docs.pytest.org/en/latest/how-to/fixtures.html#fixtures-can-introspect-the-requesting-test-context